refactor(cli): Add Installer helper to encapsulate publishing logic
Extracted installing operations into a dedicated Installer struct to reduce parameter passing and improve code organization.
