+++ src/command.rs
@@ -13,19 +13,14 @@
 // limitations under the License.
 
 use crate::{
-    cache::{Cache, Entry as CacheEntry},
     credentials::Credentials,
     lock::Lockfile,
-    manifest::{Dependency, MANIFEST_FILE, Manifest, PackageManifest},
-    package::{Package, PackageName, PackageStore, PackageType},
+    manifest::{Dependency, MANIFEST_FILE, Manifest, ManifestType, PackageManifest},
+    package::{PackageName, PackageStore, PackageType},
     registry::{Artifactory, RegistryUri},
     resolver,
 };
 
-use crate::lock::LOCKFILE;
-use crate::manifest::{
-    DependencyManifest, ManifestType, RemoteDependencyManifest,
-};
 use crate::resolver::DependencySource;
 use miette::{Context as _, IntoDiagnostic, bail, ensure, miette};
 use semver::{Version, VersionReq};
@@ -38,6 +33,7 @@ use tokio::{
     fs,
     io::{self, AsyncBufReadExt, BufReader},
 };
+use crate::operations::installer::Installer;
 use crate::operations::publisher::Publisher;
 
 const INITIAL_VERSION: Version = Version::new(0, 1, 0);
@@ -416,220 +412,8 @@ async fn publish_package(
 /// * `preserve_mtime` - If true, local dependencies preserve their modification time
 pub async fn install(preserve_mtime: bool) -> miette::Result<()> {
     let manifest = Manifest::read().await?;
-
-    match manifest.manifest_type {
-        ManifestType::Package => {
-            let lockfile = Lockfile::read_or_default().await?;
-            let store = PackageStore::current().await?;
-            let cache = Cache::open().await?;
-            let current_path = env::current_dir()
-                .into_diagnostic()
-                .wrap_err("current dir could not be retrieved")?;
-
-            install_package(
-                preserve_mtime,
-                &manifest,
-                &lockfile,
-                &store,
-                &current_path,
-                &cache,
-            )
-            .await
-        }
-        ManifestType::Workspace => install_workspace(preserve_mtime, &manifest).await,
-    }
-}
-
-/// Installs dependencies for a workspace (not yet implemented)
-async fn install_workspace(preserve_mtime: bool, manifest: &Manifest) -> miette::Result<()> {
-    let root_path = env::current_dir()
-        .into_diagnostic()
-        .wrap_err("current dir could not be retrieved")?;
-    let workspace = manifest.workspace.as_ref().ok_or_else(|| {
-        miette!(
-            "buffers install for workspaces executed on manifest that does not define a workspace."
-        )
-    })?;
-    let packages = workspace.resolve_members(root_path)?;
-    tracing::info!(
-        ":: workspace found. running install for {} packages in workspace",
-        packages.len()
-    );
-
-    let cache = Cache::open().await?;
-
-    for package in packages {
-        let canonical_name = fs::canonicalize(&package).await.into_diagnostic()?;
-        let pkg_manifest = Manifest::try_read_from(package.join(MANIFEST_FILE)).await?;
-        let pkg_lockfile = Lockfile::read_from_or_default(package.join(LOCKFILE)).await?;
-        let store = PackageStore::open(&package).await?;
-
-        tracing::info!(
-            ":: running install for package: {}",
-            canonical_name.to_str().unwrap()
-        );
-        install_package(
-            preserve_mtime,
-            &pkg_manifest,
-            &pkg_lockfile,
-            &store,
-            &package,
-            &cache,
-        )
-        .await?
-    }
-
-    Ok(())
-}
-
-/// Installs dependencies of a package
-///
-/// if [preserve_mtime] is true, local dependencies will keep their modification time
-async fn install_package(
-    preserve_mtime: bool,
-    manifest: &Manifest,
-    lockfile: &Lockfile,
-    store: &PackageStore,
-    package_path: &PathBuf,
-    cache: &Cache,
-) -> miette::Result<()> {
-    let credentials = Credentials::load().await?;
-
-    store.clear().await?;
-
-    if let Some(ref pkg) = manifest.package {
-        store.populate(pkg).await?;
-
-        tracing::info!(":: installed {}@{}", pkg.name, pkg.version);
-    }
-
-    let graph_v2 = resolver::DependencyGraph::build(manifest, package_path, &credentials).await?;
-    let dependencies = graph_v2.ordered_dependencies()?;
-
-    let mut locked = Vec::new();
-
-    for dependency_node in dependencies {
-        // Iterate through the dependencies in order and install them
-        let package = match dependency_node.node.source {
-            // corresponds to process_local_dependency in resolver.rs:
-            // Key logic is: Read manifest, create directory structures via PackageStore::open, initialize package via store.release
-            DependencySource::Local { path } => {
-                // For local dependencies, create a store at the dependency path and release it
-                let dep_manifest = Manifest::try_read_from(path.join(MANIFEST_FILE)).await?;
-                let dep_store = PackageStore::open(&path).await?;
-                dep_store.release(&dep_manifest, preserve_mtime).await?
-            }
-
-            // corresponds to process_remote_dependency in resolver.rs:
-            DependencySource::Remote {
-                repository,
-                registry,
-            } => {
-                let package_name = &dependency_node.node.name;
-                let version = &dependency_node.node.version;
-
-                // Try to use cached package if available in lockfile
-                let mut resolved_package = None;
-                if let Some(locked_entry) = lockfile.get(package_name) {
-                    // Validate registry consistency
-                    ensure!(
-                        registry == locked_entry.registry,
-                        "registry mismatch for {}: manifest specifies {} but lockfile requires {}",
-                        package_name,
-                        registry,
-                        locked_entry.registry
-                    );
-
-                    // Try to retrieve from cache if version matches lockfile
-                    if version.matches(&locked_entry.version) {
-                        if let Some(cached_pkg) = cache.get(locked_entry.into()).await? {
-                            // Validate the cached package digest
-                            locked_entry.validate(&cached_pkg)?;
-
-                            tracing::debug!(
-                                ":: using cached package for {}@{}",
-                                package_name,
-                                cached_pkg.version()
-                            );
-
-                            resolved_package = Some(cached_pkg);
-                        }
-                    }
-                }
-
-                // Download from registry if not cached
-                let resolved_package = match resolved_package {
-                    Some(pkg) => pkg,
-                    None => {
-                        let artifactory = Artifactory::new(registry.clone(), &credentials)
-                            .wrap_err(miette!("failed to initialize registry {}", registry))?;
-
-                        download_and_cache(
-                            &artifactory,
-                            package_name,
-                            &registry,
-                            &repository,
-                            version,
-                            cache,
-                        )
-                        .await?
-                    }
-                };
-
-                // Add to new lockfile
-                let dependants_count = graph_v2.dependants_count_of(package_name);
-                locked.push(resolved_package.lock(registry, repository, dependants_count));
-
-                resolved_package
-            }
-        };
-
-        store
-            .unpack(&package)
-            .await
-            .wrap_err(miette!("failed to unpack package {}", &package.name()))?;
-
-        tracing::info!(
-            ":: installed {}@{}",
-            dependency_node.name,
-            package.version()
-        );
-    }
-
-    // Write lockfile
-    Lockfile::from_iter(locked.into_iter())
-        .write(package_path)
-        .await
-}
-
-/// Downloads a package from the registry and caches it
-async fn download_and_cache(
-    artifactory: &Artifactory,
-    package_name: &PackageName,
-    registry: &RegistryUri,
-    repository: &str,
-    version: &VersionReq,
-    cache: &Cache,
-) -> miette::Result<Package> {
-    let dependency = Dependency {
-        package: package_name.clone(),
-        manifest: DependencyManifest::Remote(RemoteDependencyManifest {
-            registry: registry.clone(),
-            repository: repository.to_string(),
-            version: version.clone(),
-        }),
-    };
-
-    let downloaded_package = artifactory.download(dependency).await?;
-
-    // Cache the downloaded package for future installs
-    let cache_key = CacheEntry::from(&downloaded_package);
-    cache
-        .put(cache_key, downloaded_package.tgz.clone())
-        .await
-        .ok();
-
-    Ok(downloaded_package)
+    let installer = Installer::new(preserve_mtime).await?;
+    installer.install(&manifest).await
 }
 
 /// Uninstalls dependencies
